surjective generator works better.
The surjective generator now works properly when seeded with a mapping that has
a branch point that maps to something other than the branch point on the
codomain triod.

The problem occured because the point preceeding the key point of the domain
triod was already mapping to the endpoint of the codomain triod. In the code,
this meant that X == 1 and Y == 0, causing the code to go down the 'standard
ajacent point' pathway. This meant that the first result was correct; however
the completions would then continue. to the previous ajacent point, and so on,
which was what was causing the error.