diagnosed the surjectivity problem
The problem with the surjective generator only seems to occur when
the mappings are partially complete before being submitted to the
surjective generator.

The problem has not actually been resolved yet. That still needs to be done.

Also implemented a log function to clean up sections of the code which print.