Fixed surjective generator working with partial completions
A problem existed where the generator would complete the partial mapping using one
end point map, but then would continue using another. This had the effect of both
wasting time as well as causing problems. It's odd that the surjective generator
did not catch this before returning a mapping. That will have to be looked into.

Anyway, it seems to work now.